adds new wrapper for 3dView by splitting out an abstract 3dview wrapper from perspective, axonometric and regular 3dview wrappers.
